Added menus to the context menu
