+++ webstack/apps/webapp/src/app/pages/board/layers/ui/components/Menus/Menus/ApplicationsMenu.tsx
